requirements.txt
litellm>=1.0.0
pandas>=2.0.0
datasets>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
